pyproject.toml
[tool.poetry]
name = "oauth-sandbox"
version = "0.1.0"
description = ""
authors = ["Vincent Ging Ho Yim <28189684+cenviity@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests-oauthlib = "^1.3.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
